+++ gdsfactory/tests/test_components.py
@@ -1,7 +1,7 @@
 import pytest
 from pytest_regressions.data_regression import DataRegressionFixture
 
-from gdsfactory.components import component_factories
+from gdsfactory.components import cells
 from gdsfactory.difftest import difftest
 
 skip_test = {
@@ -12,27 +12,27 @@ skip_test = {
     "mzi_arm",
 }
 
-components_to_test = set(component_factories.keys()) - skip_test
+cells_to_test = set(cells.keys()) - skip_test
 
 
-@pytest.fixture(params=components_to_test, scope="function")
+@pytest.fixture(params=cells_to_test, scope="function")
 def component_name(request) -> str:
     return request.param
 
 
 def test_gds(component_name: str) -> None:
     """Avoid regressions in GDS geometry shapes and layers."""
-    component = component_factories[component_name]()
+    component = cells[component_name]()
     difftest(component, test_name=component_name)
 
 
 def test_settings(component_name: str, data_regression: DataRegressionFixture) -> None:
     """Avoid regressions when exporting settings."""
-    component = component_factories[component_name]()
+    component = cells[component_name]()
     data_regression.check(component.to_dict())
 
 
 def test_assert_ports_on_grid(component_name: str):
     """Ensure ports are on grid."""
-    component = component_factories[component_name]()
+    component = cells[component_name]()
     component.assert_ports_on_grid()