+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [4.7.3]
 
-- add `has_routing_info` to CrossSection to ensure it has routing information
+- add `has_routing_info` to [CrossSection](CrossSection) to ensure it has routing information
 - rename cross_section_factory to cross_sections
 - rename component_factory to cells